+++ CHANGELOG.md
@@ -13,3 +13,10 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 ## 0.1.7
 
 - feat: get or put miscellaneous data in cart
+
+## 0.1.8
+
+- feat: allow `shipping` as a condition type
+- fix: add(...): condition return type potential return type error
+- fix: apply(...): use condition return type and return applied condition(s)
+- feat: apply(...): replace already existing condition if exists